+++ docs/changelog.md
@@ -35,6 +35,8 @@ What's changed since pre-release v1.48.0-B0228:
 - General improvements:
   - Added support for resolving filtered subnet IDs from existing virtual networks during Bicep expansion.
     [#2159](https://github.com/Azure/PSRule.Rules.Azure/issues/2159)
+  - Added support for the Bicep `distinct` and `like` functions.
+    [#3905](https://github.com/Azure/PSRule.Rules.Azure/issues/3905)
 - Updated rules:
   - Container Registry:
     - Deprecated `Azure.ACR.GeoReplica` because ACR zone redundancy is automatic in supported regions.